requirements.txt
Django
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
python-dotenv
pytz
sqlparse
asgiref

# Optional for testing/dev
coverage
pytest
pytest-django
black
flake8

# ai
groq
